refactor(ui): move graph downsampling next to axis helpers
Keeps result_graph.go under the 200 LOC cap and covers long-run
downsampling with a width + reveal-reflow test.
