+++ internal/ui/result_graph.go
@@ -145,26 +145,8 @@ func RenderResultGraph(perSec []metrics.PerSecond, width, chartH, visible int, t
 	return b.String()
 }
 
-// bucketSamples folds perSec into ceil(len/secPerCell) buckets: mean RawWPM,
-// summed Errors. Sec keeps each bucket's starting second.
-func bucketSamples(perSec []metrics.PerSecond, secPerCell int) []metrics.PerSecond {
-	out := make([]metrics.PerSecond, 0, (len(perSec)+secPerCell-1)/secPerCell)
-	for i := 0; i < len(perSec); i += secPerCell {
-		end := i + secPerCell
-		if end > len(perSec) {
-			end = len(perSec)
-		}
-		var b metrics.PerSecond
-		b.Sec = perSec[i].Sec
-		for _, ps := range perSec[i:end] {
-			b.RawWPM += ps.RawWPM
-			b.Errors += ps.Errors
-		}
-		b.RawWPM /= float64(end - i)
-		out = append(out, b)
-	}
-	return out
-}
+// bucketSamples (downsampling) and the axis/scale helpers live in
+// result_graph_axes.go.
 
 // drawSeg connects (x0,y0)-(x1,y1) on the dot grid by sampling densely enough to
 // fill both the vertical and horizontal spans, so flat segments render solid and
